remove unused fields
